feat(web): redesign OSS landing + share post-signin redirect helpers
- Replace the OSS landing copy with the agentic-era headline + body, drop the eyebrow and old tagline, render the Munin logo above the title, and extend the bone background to cover the full viewport.
- Remove the OSS footer and the OSS-side privacy/terms routes (+ their legal markdown). getmunin.com legal pages will live elsewhere; cloud retains its own app.getmunin.com policies.
- Drop now-unused deps (react-markdown, @tailwindcss/typography) and the typography plugin from the OSS web Tailwind config.
- Move safeRedirect + resumeOauthAuthorizeUrl into @getmunin/dashboard-pages (new auth/post-signin-redirect.ts) and consume them from the OSS auth pages. Cloud auth pages keep a local copy until the next dashboard-pages release.
