--- apps/web/components/Footer.tsx
@@ -1,53 +1,0 @@
-'use client';
-
-import { Link, usePathname } from '@/i18n/navigation';
-import { useTranslations } from 'next-intl';
-
-export function Footer() {
-  const pathname = usePathname();
-  const t = useTranslations('footer');
-
-  if (pathname?.startsWith('/dashboard')) return null;
-  if (pathname === '/login' || pathname === '/signup') return null;
-  if (pathname?.startsWith('/accept-invite')) return null;
-
-  const isDocs = pathname?.startsWith('/docs');
-
-  return (
-    <footer
-      className={
-        isDocs
-          ? 'border-t-[0.5px] border-rule-soft px-14 py-5 text-xs text-ink-mute max-[880px]:px-5 dark:border-rule-on-dark'
-          : 'border-t-[0.5px] border-rule-soft px-6 py-5 text-xs text-ink-mute dark:border-rule-on-dark'
-      }
-    >
-      <nav
-        className={
-          isDocs
-            ? 'flex flex-wrap items-center justify-between gap-4'
-            : 'mx-auto flex max-w-5xl flex-wrap items-center justify-between gap-4'
-        }
-      >
-        <span className="font-mono text-[10px] uppercase tracking-eyebrow">
-          © {new Date().getFullYear()} Munin
-        </span>
-        <div className="flex items-center gap-6 font-mono text-[10px] uppercase tracking-eyebrow">
-          <Link href="/privacy" className="hover:text-ink dark:hover:text-foreground">
-            {t('privacy')}
-          </Link>
-          <Link href="/terms" className="hover:text-ink dark:hover:text-foreground">
-            {t('terms')}
-          </Link>
-          <a
-            href="https://github.com/getmunin/munin"
-            className="hover:text-ink dark:hover:text-foreground"
-            rel="noopener noreferrer"
-            target="_blank"
-          >
-            GitHub
-          </a>
-        </div>
-      </nav>
-    </footer>
-  );
-}
