+++ apps/web/package.json
@@ -30,7 +30,6 @@
     "next-themes": "^0.4.6",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
-    "react-markdown": "^9.1.0",
     "@sentry/nextjs": "^10.51.0",
     "shadcn": "^4.5.0",
     "sonner": "^2.0.7",
@@ -41,7 +40,6 @@
   "devDependencies": {
     "@getmunin/eslint-config": "workspace:*",
     "@getmunin/tsconfig": "workspace:*",
-    "@tailwindcss/typography": "^0.5.19",
     "@types/node": "^22.7.0",
     "@types/react": "^19.0.0",
     "@types/react-dom": "^19.0.0",
