--- apps/web/public/legal/privacy.md
@@ -1,77 +1,0 @@
-# Privacy Policy — getmunin.com
-
-_Last updated: 2026-04-30_
-
-This policy covers **getmunin.com** — the public marketing site, the open-source documentation, and the GitHub Packages registry that serves the `@getmunin/*` libraries. It is published by Apps AS, registered at Vulkan 16, 0178 Oslo, Norway ("we", "us"). For the purposes of the EU GDPR we are the **data controller** for the personal data described below.
-
-If you run Munin as a **self-hosted** instance under your own control, this policy does **not** cover the data your instance processes. You are the data controller for that data; this policy only governs data that flows through getmunin.com itself.
-
-If you use **Munin Cloud** (https://app.getmunin.com), see the separate [Munin Cloud Privacy Policy](https://app.getmunin.com/privacy).
-
-## 1. What we collect on getmunin.com
-
-We collect the minimum needed to run the site and OSS distribution.
-
-| Category | Data | Source |
-| --- | --- | --- |
-| Server logs | IP address, user agent, request path, timestamp, response status | Automatic on every HTTP request |
-| Support contacts | Name, email, message body | When you email us at hello@getmunin.com or security@getmunin.com |
-| OSS package distribution | GitHub username and download metadata | Provided by GitHub when you install `@getmunin/*` from GitHub Packages |
-| Cookies | A locale preference cookie (`NEXT_LOCALE`) | Set when you switch language |
-
-We do **not** run third-party analytics, ad pixels, or session-replay tools on getmunin.com. The locale cookie is the only cookie we set; it is strictly necessary for the language switcher and is exempt from consent under the ePrivacy Directive.
-
-## 2. Why we process this data
-
-| Purpose | Legal basis (GDPR Art. 6) |
-| --- | --- |
-| Operating and securing the site (logs, abuse prevention) | Legitimate interest (Art. 6(1)(f)) |
-| Replying to support and security reports you send us | Legitimate interest, and performance of measures taken at your request (Art. 6(1)(b),(f)) |
-| Distributing the open-source code via GitHub Packages | Legitimate interest in maintaining the OSS project |
-| Remembering your language preference | Legitimate interest; strictly necessary functionality |
-
-We do not sell personal data and we do not use it to train AI models.
-
-## 3. Retention
-
-- **HTTP server logs:** up to 30 days, then deleted.
-- **Support emails:** for as long as is reasonably necessary to handle the conversation and any follow-up; typically 24 months.
-- **GitHub Packages downloads:** retained by GitHub under [GitHub's privacy policy](https://docs.github.com/site-policy/privacy-policies/github-general-privacy-statement); we receive aggregated counters only.
-- **Locale cookie:** 12 months from your last visit.
-
-## 4. Who we share it with
-
-We use a small number of processors to run getmunin.com:
-
-| Processor | What it processes | Where |
-| --- | --- | --- |
-| The hosting provider for getmunin.com | Server logs, request bodies | EU |
-| GitHub, Inc. | OSS source, releases, package downloads, issue tracker | US, with SCCs |
-| Our email provider (used for support correspondence) | Inbound and outbound mail | EU/US, with SCCs where applicable |
-
-We do not transfer personal data outside these processors. We do not share data with advertisers, brokers, or analytics partners.
-
-## 5. Your rights (EU/EEA, UK, California)
-
-You have the right to access, correct, delete, restrict, port, and object to our processing of your personal data. EU/EEA residents may also lodge a complaint with their supervisory authority. California residents have the rights under the CCPA/CPRA, including the right to know and delete; we do not sell or share personal information as those terms are defined under California law.
-
-To exercise these rights, email **privacy@getmunin.com**. We will respond within 30 days.
-
-## 6. Security
-
-- All getmunin.com traffic is served over TLS.
-- We follow the practices in our [Security Policy](https://github.com/getmunin/munin/blob/main/SECURITY.md). Vulnerabilities can be reported to **security@getmunin.com**; we acknowledge within 72 hours and provide a remediation timeline within 7 days for confirmed issues.
-
-## 7. Children
-
-getmunin.com is not directed at children under 16 and we do not knowingly collect their personal data. If you believe we have, contact privacy@getmunin.com and we will delete it.
-
-## 8. Changes
-
-We will post material changes on this page and update the "Last updated" date. For substantive changes we will, where feasible, also notify recent support contacts by email.
-
-## 9. Contact
-
-Apps AS
-Vulkan 16, 0178 Oslo, Norway
-privacy@getmunin.com
