--- apps/web/public/legal/terms.md
@@ -1,59 +1,0 @@
-# Terms of Use — getmunin.com
-
-_Last updated: 2026-04-30_
-
-These terms govern your use of **getmunin.com** — the public marketing site, the open-source documentation, the public suggestions board, and the GitHub Packages registry that serves the `@getmunin/*` libraries. They are published by Apps AS, registered at Vulkan 16, 0178 Oslo, Norway ("we", "us").
-
-For the **Munin Cloud** hosted SaaS (https://app.getmunin.com), separate [Munin Cloud Terms](https://app.getmunin.com/terms) apply.
-
-## 1. The open-source software
-
-The Munin source code and the `@getmunin/*` libraries are licensed under the **MIT License** (see [LICENSE](https://github.com/getmunin/munin/blob/main/LICENSE)). The MIT License grants you broad rights and disclaims warranties and liability — read it; it governs your use of the code.
-
-When you use the code, including via `pnpm add` from GitHub Packages, you accept the MIT License's terms. These terms of use do **not** override the MIT License with respect to the code itself.
-
-## 2. The website
-
-You may browse getmunin.com, read the documentation, and link to it. You agree not to:
-
-- attempt to disrupt the site (denial of service, automated scraping at a scale that degrades the service for others, exploiting vulnerabilities outside the bounds of our [Security Policy](https://github.com/getmunin/munin/blob/main/SECURITY.md));
-- use the site to distribute malware, phishing content, or unlawful material;
-- impersonate us or misrepresent your relationship with the project.
-
-We may rate-limit or block traffic that violates these terms or threatens the service.
-
-## 3. Suggestions and other contributions
-
-The public suggestions board at https://getmunin.com/suggestions and the GitHub issue tracker accept user-generated content. By posting:
-
-- you grant us a non-exclusive, royalty-free, worldwide, perpetual licence to display, reproduce, and act upon your post within the project (including publishing aggregate vote counts, quoting it in changelogs, and shipping a feature in response);
-- you confirm the content is yours to share and does not violate anyone else's rights;
-- you accept that posts may be edited or removed if they are off-topic, abusive, infringing, or otherwise inappropriate.
-
-We don't use posted content to train AI models.
-
-## 4. Trademarks
-
-"Munin" and the Munin logo are our trademarks. The MIT License grants you rights to the code, not to our brand. Don't use the name or logo in a way that suggests endorsement or affiliation we haven't agreed to. Forks must use a different name and logo if distributed publicly.
-
-## 5. No warranty (the website)
-
-getmunin.com is provided **"as is"**, without warranty of any kind. We don't warrant that the site will be uninterrupted, error-free, or available in your region. The MIT License's warranty disclaimer governs the code itself.
-
-## 6. Limitation of liability
-
-To the maximum extent permitted by law, neither we nor our contributors will be liable for indirect, incidental, special, consequential, or punitive damages, or for lost profits, data, or goodwill, arising out of your use of getmunin.com or the open-source code. Our total liability arising out of these terms or your use of the site, in aggregate, will not exceed €100. Nothing in these terms limits liability that cannot be limited under applicable law (including liability for fraud, gross negligence, or death/personal injury caused by negligence).
-
-## 7. Changes
-
-We may update these terms. The current version is always at this URL with a "Last updated" date. Material changes will be announced in the changelog or repository. Continued use of the site after a change constitutes acceptance.
-
-## 8. Governing law
-
-These terms are governed by the laws of Norway, without regard to its conflict-of-laws rules. Disputes will be brought before the courts of Oslo, Norway, except where mandatory consumer-protection law of your country of residence requires otherwise.
-
-## 9. Contact
-
-Apps AS
-Vulkan 16, 0178 Oslo, Norway
-hello@getmunin.com
